Modify config clock frequency for 6S speed grade
Boards that implement the 6S speed grade device have a 200MHz config
clock frequency, not 250Mhz like 4S and 5S speed grade devices. Some of
the demos rely on the frequency of this clock to be defined accurately.
This commit implements script changes that will edit the affected source
files before they are used to create the system that rely on an accurate
frequency.

Signed-off-by: Rod Frazer <rod.frazer@intel.com>