Change emac completion detection to interrupt
Changing the emac software to monitor for an interrupt assertion
triggered from the last RX descriptor to indicate that the data movement
is complete. The original method of waiting for the DMA to indicate it
had exhausted all TX and RX descriptors was not working reliably on the
Critical Link board.

Signed-off-by: Rod Frazer <rod.frazer@intel.com>