Add synchronization before reading system counter
The initial code implemetation applied loose synchronization before
reading the system counter and in some situations it was too loose and
would allow the system counter to be read too early or too late to be
useful as a timing measurement. This commit attempts to implement a very
agressive synchronization prior to any system counter read that might be
used for a timing calculation. The agressive synchronization includes
both a dsb and an isb instruction prior to reading the system counter.

In situations where the system counter was just being read for
non-timing critical purposes, all synchronization has been removed.

Signed-off-by: Rod Frazer <rod.frazer@intel.com>