Update to QPDS 24.3.1 tools release
This is a significant update to all hardware and software objects within
this repo as they are all updated to build under the QPDS 24.3.1 tools
environment.

The Quartus 24.3.1 build 102 tools release contains a defect that is
resolved by applying patch 1.14. This is required by this release in
order to properly build the hardware examples.

Some of the more notable changes are listed below:

Agilex 5 SoC component is upgraded to version 6.0.0

HPS EMIF component is upgraded to version 2.0.0

Configuration Clock component is upgraded to version 19.1.5

Reset Release IP component is upgraded to version 1.0.1

u-boot-socfpga, arm-trusted-firmware and linux-socfpga are updated to
the latest tag, QPDS24.3.1_REL_GSRD_PR

toybox is updated to the latest tag, 0.8.12

Disabled all advanced calibration options for the HPS EMIF controller in
all of the designs. This shortens the boot time consumed by EMIF
calibration.

Added the clusterpwrstat_el1 to the cache_regs.c application which
displays how much L3 cache is enabled in the cluster.

In the 24.3.1 tools, the aXmmusecsid value was changed from 1'b0 to 1'b1
and this prevents cache coherent transactions from interacting with the
cache subsystem properly when they pass through the F2H bridge. We have
implemented a script that sets that value back to 1'b0 and allows the
cache coherent demo to operate properly again.

Signed-off-by: Rod Frazer <rod.frazer@altera.com>