Update to QPDS 25.1 tools release
This is a significant update to all hardware and software objects within
this repo as they are all updated to build under the QPDS 2.1 tools
environment.

Some of the more notable changes are listed below:

The following IP cores are updated in the 25.1 tools:

    altera_avalon_sysid_qsys 19.1.8
    altera_i2cslave_to_avlmm_bridge  19.1.6
    altera_msgdma 19.2.5
    altera_s10_configuration_clock 19.1.6
    emif_io96b_hps 3.0.0
    intel_agilex_5_soc 7.0.0
    intel_lw_uart lw_uart 1.0.8
    intel_niosv_c 4.0.0
    intel_onchip_memory 1.4.10

The 25.1 tools introduce a new Quartus project setting that we need to
enable the HPS to acquire ownership of the SDM QSPI flash device:

    set_global_assignment -name QSPI_OWNERSHIP HPS

The Linux kernel build needs to add the following flag to the make
invocation:

    DTC_FLAGS=-@

The Agilex 5 devicetree in the Linux kernel source tree moves the FPGA
region node:

    FROM: target-path = /soc@0/base_fpga_region;
      TO: target-path = /fpga-region;

Some additional devicetree syntax needed to be modified for proper
operation as well. The placement of these properties needed to be
relocated:

    /delete-node/ voltage;
    /delete-node/ temperature;

In 25.1 the H2F bridge hardware design has an HPS IO hash mismatch with
respect to all of the other designs. By placing an AXI4 bridge in front
of the H2F bridge and connecting all of the subordinates to the AXI4
bridge, the HPS IO hash seems to consistently match the hash of the rest
of the designs.

u-boot-socfpga, arm-trusted-firmware and linux-socfpga are updated to
the latest tag, QPDS25.1_REL_GSRD_PR

toybox remains at the latest tag, 0.8.12

In the 6.12-lts kernel, the iommu kernel driver framework does not allow
peripherals to bypass translation as easily as previous kernel releases.
There are a number of ways to work around this, we chose to add
devicetree properties to define iommu-addresses for our reserved memory
buffer so that the kernel will provide a 1:1 physical to logical mapping
for the buffer span. This approach is a little messy as overlays are
added and removed, but it seems to work reliably and consistently.

The devicetree overlays for the f2h_bridge and emac examples are updated
with this property.

The 6.12-lts kernel introduces a new kernel config for the Designware
I2C driver, CONFIG_I2C_DESIGNWARE_CORE, which has been added to our
kernel config.

Signed-off-by: Rod Frazer <rod.frazer@altera.com>